create NamedArgsMixin in rv.py
